requirements.txt
flask
transformers
torch
pymupdf 
python-docx 
openpyxl
fitz 
cohere 
PyPDF2

